Started the task manager to thread execution

+++ src/Maths/Vector2f.java
@@ -3,45 +3,45 @@ package Maths;
 /**
  * Created by steppers on 28/10/15.
  */
-public class Vector2i {
+public class Vector2f {
 
-    public int x, y;
+    public float x, y;
 
-    public Vector2i(int x, int y)
+    public Vector2f(float x, float y)
     {
         this.x = x;
         this.y = y;
     }
 
-    public Vector2i add(int x, int y)
+    public Vector2f add(float x, float y)
     {
         this.x += x;
         this.y += x;
         return this;
     }
 
-    public Vector2i add(Vector2i v)
+    public Vector2f add(Vector2f v)
     {
         this.x += v.x;
         this.y += v.y;
         return this;
     }
 
-    public Vector2i sub(int x, int y)
+    public Vector2f sub(float x, float y)
     {
         this.x -= x;
         this.y -= x;
         return this;
     }
 
-    public Vector2i sub(Vector2i v)
+    public Vector2f sub(Vector2f v)
     {
         this.x -= v.x;
         this.y -= v.y;
         return this;
     }
 
-    public Vector2i mul(int m)
+    public Vector2f mul(float m)
     {
         this.x *= m;
         this.y *= m;
